+++ tools/decode-status.py
@@ -196,7 +196,8 @@ a_setoption = [[
     "(Pressure) Switch between mmHg (0) or inHg (1) when SO24 1",
     "(MQTT) MQTT clean session (0 = default) or persistent session (1)",
     "(GUI) Disable display of GUI module name (1)",
-    "","","",""
+    "(Wifi) Wait 1 second for wifi connection solving some FRITZ!Box modem issues (1)",
+    "","",""
     ],[
     "","","","",
     "","","","",